pyproject.toml
[project]
name = "moroccan-law-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.5",
    "dotenv>=0.9.9",
    "langchain>=1.2.3",
    "langchain-neo4j>=0.7.0",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "streamlit-arabic-support-wrapper>=1.1",
    "tavily-python>=0.7.17",
]
